tutorials/airflow_on_flyte/requirements.txt
apache-airflow-providers-slack
flytekitplugins-airflow

